requirements.txt
streamlit==1.30.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
scikit-surprise==1.1.3
matplotlib==3.8.0
seaborn==0.13.0
